+++ src/Main2/cart_functions/CartManager.java
@@ -1,97 +1,53 @@
 package cart_functions;
-import browse.GetUserDictGateway.GetUserDictGatewayInterface;
-import browse.SaveUserDictGateway.SaveUserDictGatewayInterface;
-import login_functions.SaveUserGatewayInterface;
-import post.Post;
-import product.*;
+
+import options.buy_functions.BuyController;
+import product.Product;
 import user.User;
 
 import java.io.IOException;
-import java.util.HashMap;
+import java.util.*;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import java.util.List;
 
-/**
- * A use case class that adds items to the user's cart
- */
+import options.follow_users.UserUseCase;
+
 public class CartManager {
-    SaveProductGatewayInterface saveProductChangesInterface;
-    SaveUserGatewayInterface saveUserGatewayInterface;
-    GetUserDictGatewayInterface getUserDictGatewayInterface;
-    SaveUserDictGatewayInterface saveUserDictGatewayInterface;
 
-    public CartManager(SaveProductGatewayInterface saveProductChangesInterface, GetUserDictGatewayInterface getUserDictGatewayInterface, SaveUserDictGatewayInterface saveUserDictGatewayInterface){
+    public void buyCart(User user) throws Exception {
+        List<String> cartIDs = new ArrayList<>();
+        UserUseCase userUseCase = new UserUseCase(user);
 
-        this.saveProductChangesInterface = saveProductChangesInterface;
-        this.saveUserDictGatewayInterface = saveUserDictGatewayInterface;
-        this.getUserDictGatewayInterface = getUserDictGatewayInterface;
-    }
-
-    public CartManager(SaveUserGatewayInterface saveUserGatewayInterface ){
-        this.saveUserGatewayInterface = saveUserGatewayInterface;
-    }
-
-    public CartManager(){};
-
-    /**
-     *
-     * @param item The product that we want to add to the cart
-     * @param user The user whom we want to add the product to their cart
-     * @return true or false if the product was or was not added to the user's cart
-     */
-
-    public boolean addToCart(Product item, User user){
-        List<Product> userCart = user.getShoppingCart();
-        if (item.getQuantity() > 0){
-            userCart.add(item);
-            return true;
+        for (Product i:userUseCase.userShoppingCart()) {
+            cartIDs.add(i.getId());
         }
-        return false;
-    }
-
-    /**
-     * Updates the quantity of the cart when the user buys
-     * @param user the user whom we want to update their cart
-     */
-
-    public void updateProductQuantity(User user) throws IOException, ClassNotFoundException {
-        List<Product> userCart = user.getShoppingCart();
-        HashMap <String, User> userDict = getUserDictGatewayInterface.getUserDict();
-        for (Product prod: userCart){
-            GetProductGatewayInterface getProductGatewayInterface = new GetProductGateway();
-            ProductUseCase prodUseCase = new ProductUseCase(getProductGatewayInterface);
-            prodUseCase.decreaseQuantity(prod, 1);
-            // Todo: delete the product with the old information from the file
-            saveProductChangesInterface.save(prod.getId(),prod);
-            //updating product in post serilzation in the userDict
-            User productUser = userDict.get(prod.getUsername());
-            List postList = (List) productUser.getListPosts();
-            for (int c = 0; c < postList.size(); c++){
-                Post post = (Post) postList.get(c);
-                if (prod.getId().equals(post.getProduct().getId())){
-                    Product productPost = post.getProduct();
-                    prodUseCase.decreaseQuantity(productPost, 1);
-                    post.setProduct(productPost);
-                    postList.set(c, post);
-                }
-            }
-            productUser.setListPosts(postList);
-            userDict.put(prod.getUsername(),productUser);
+        for (Product i:userUseCase.userShoppingCart()){
+            this.buyItem(user, i);
         }
-        saveUserDictGatewayInterface.saveUserDict(userDict);
     }
 
-    /**
-     *
-     * @param user The user whom we will empty their cart
-     * @throws IOException
-     * @throws ClassNotFoundException
-     */
+    public void addToCart(User user, Product product){
+        List<Product> productList = new ArrayList<>();
+        productList.add(product);
 
-    public void emptyCart(User user) throws IOException, ClassNotFoundException {
-        List<Product> userCart = user.getShoppingCart();
-        userCart.clear();
-        user.setShoppingCart(userCart);
-        saveUserGatewayInterface.saveUser(user.getUsername(), user);
+        user.setShoppingCart(Stream.concat(user.getShoppingCart().stream(),
+                productList.stream()).collect(Collectors.toList()));
     }
+
+    public void updateProductQuantity(Product product){
+        product.setQuantity(product.getQuantity() - 1);
+    }
+
+    public void buyItem(User user, Product product) throws IOException, ClassNotFoundException {
+        BuyController buyController = new BuyController();
+        List<String> cartIDs = new ArrayList<>();
+        UserUseCase userUseCase = new UserUseCase(user);
+
+        for (Product i:userUseCase.userShoppingCart()) {
+            cartIDs.add(i.getId());
+        }
+        buyController.allowBuy(user, cartIDs, cartIDs.indexOf(product.getId()));
+        this.updateProductQuantity(product);
+    }
+
 }
-
